Introduce `JavaScriptBuilderElementTestsBase`. pt.1.
